pyproject.toml
[tool.poetry]
name = "rtpengine-client"
version = "0.1.0"
description = ""
authors = ["raushanraja <vraushanraja@gmail.com>"]
readme = "README.md"
packages = [{include = "rtpengine_client"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
bencode-py = "^4.0.0"
aiohttp = "^3.8.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
